+++ config/test.exs
@@ -16,7 +16,7 @@ config :plausible, Plausible.Mailer, adapter: Bamboo.TestAdapter
 
 config :plausible,
   paddle_api: Plausible.PaddleApi.Mock,
-  google_api: Plausible.Google.Api.Mock
+  google_api: Plausible.Google.API.Mock
 
 config :bamboo, :refute_timeout, 10