+++ lib/plausible_web/controllers/api/external_stats_controller.ex
@@ -12,7 +12,8 @@ defmodule PlausibleWeb.Api.ExternalStatsController do
     :bounce_rate,
     :visit_duration,
     :events,
-    :conversion_rate
+    :conversion_rate,
+    :time_on_page
   ]
 
   @metric_mappings Enum.into(@metrics, %{}, fn metric -> {to_string(metric), metric} end)
@@ -111,7 +112,7 @@ defmodule PlausibleWeb.Api.ExternalStatsController do
       Map.get(params, "metrics", "visitors")
       |> String.split(",")
 
-    case validate_all_metrics(metrics, property, query) do
+    case validate_metrics(metrics, property, query) do
       {:error, reason} ->
         {:error, reason}
 
@@ -149,13 +150,45 @@ defmodule PlausibleWeb.Api.ExternalStatsController do
     end
   end
 
-  defp validate_all_metrics(metrics, property, query) do
+  defp validate_metrics(metrics, property, query) do
+    if length(metrics) == length(Enum.uniq(metrics)) do
+      validate_each_metric(metrics, property, query)
+    else
+      {:error, "Metrics cannot be queried multiple times."}
+    end
+  end
+
+  defp validate_each_metric(metrics, property, query) do
     Enum.reduce_while(metrics, [], fn metric, acc ->
       case validate_metric(metric, property, query) do
         {:ok, metric} -> {:cont, acc ++ [metric]}
         {:error, reason} -> {:halt, {:error, reason}}
       end
     end)
+  end
+
+  defp validate_metric("time_on_page" = metric, property, query) do
+    cond do
+      query.filters["event:goal"] ->
+        {:error, "Metric `#{metric}` cannot be queried when filtering by `event:goal`"}
+
+      query.filters["event:name"] ->
+        {:error, "Metric `#{metric}` cannot be queried when filtering by `event:name`"}
+
+      property == "event:page" ->
+        {:ok, metric}
+
+      not is_nil(property) ->
+        {:error,
+         "Metric `#{metric}` is not supported in breakdown queries (except `event:page` breakdown)"}
+
+      query.filters["event:page"] ->
+        {:ok, metric}
+
+      true ->
+        {:error,
+         "Metric `#{metric}` can only be queried in a page breakdown or with a page filter."}
+    end
   end
 
   defp validate_metric("conversion_rate" = metric, property, query) do