+++ lib/plausible_web/controllers/auth_controller.ex
@@ -697,13 +697,16 @@ defmodule PlausibleWeb.AuthController do
   end
 
   def google_auth_callback(conn, %{"error" => error, "state" => state} = params) do
-    [site_id, _redirected_to, legacy] =
+    [site_id, _redirected_to, legacy, _ga4] =
       case Jason.decode!(state) do
         [site_id, redirect_to] ->
-          [site_id, redirect_to, true]
+          [site_id, redirect_to, true, false]
 
         [site_id, redirect_to, legacy] ->
-          [site_id, redirect_to, legacy]
+          [site_id, redirect_to, legacy, false]
+
+        [site_id, redirect_to, legacy, ga4] ->
+          [site_id, redirect_to, legacy, ga4]
       end
 
     site = Repo.get(Plausible.Site, site_id)
@@ -745,15 +748,18 @@ defmodule PlausibleWeb.AuthController do
   end
 
   def google_auth_callback(conn, %{"code" => code, "state" => state}) do
-    res = Plausible.Google.HTTP.fetch_access_token(code)
+    res = Plausible.Google.API.fetch_access_token!(code)
 
-    [site_id, redirect_to, legacy] =
+    [site_id, redirect_to, legacy, ga4] =
       case Jason.decode!(state) do
         [site_id, redirect_to] ->
-          [site_id, redirect_to, true]
+          [site_id, redirect_to, true, false]
 
         [site_id, redirect_to, legacy] ->
-          [site_id, redirect_to, legacy]
+          [site_id, redirect_to, legacy, false]
+
+        [site_id, redirect_to, legacy, ga4] ->
+          [site_id, redirect_to, legacy, ga4]
       end
 
     site = Repo.get(Plausible.Site, site_id)
@@ -761,15 +767,26 @@ defmodule PlausibleWeb.AuthController do
 
     case redirect_to do
       "import" ->
-        redirect(conn,
-          external:
-            Routes.site_path(conn, :import_from_google_view_id_form, site.domain,
-              access_token: res["access_token"],
-              refresh_token: res["refresh_token"],
-              expires_at: NaiveDateTime.to_iso8601(expires_at),
-              legacy: legacy
-            )
-        )
+        if ga4 do
+          redirect(conn,
+            external:
+              Routes.google_analytics4_path(conn, :property_form, site.domain,
+                access_token: res["access_token"],
+                refresh_token: res["refresh_token"],
+                expires_at: NaiveDateTime.to_iso8601(expires_at)
+              )
+          )
+        else
+          redirect(conn,
+            external:
+              Routes.universal_analytics_path(conn, :view_id_form, site.domain,
+                access_token: res["access_token"],
+                refresh_token: res["refresh_token"],
+                expires_at: NaiveDateTime.to_iso8601(expires_at),
+                legacy: legacy
+              )
+          )
+        end
 
       _ ->
         id_token = res["id_token"]