+++ test/plausible/cache_test.exs
@@ -50,11 +50,52 @@ defmodule Plausible.CacheTest do
           assert ExampleCache.get("key", force?: true, cache_name: NonExistingCache) == nil
         end)
 
-      assert log =~ "Error retrieving key from 'NonExistingCache': :no_cache"
+      assert log =~ "Error retrieving key from 'NonExistingCache'"
     end
 
     test "cache is not ready when it doesn't exist", %{test: test} do
       refute ExampleCache.ready?(test)
+    end
+  end
+
+  describe "stats tracking" do
+    test "get affects hit rate", %{test: test} do
+      {:ok, _} = start_test_cache(test)
+      :ok = ExampleCache.merge_items([{"item1", :item1}], cache_name: test)
+      assert ExampleCache.get("item1", cache_name: test, force?: true)
+      assert {:ok, %{hit_rate: 100.0}} = Plausible.Cache.Stats.gather(test)
+      refute ExampleCache.get("item2", cache_name: test, force?: true)
+      assert {:ok, %{hit_rate: 50.0}} = Plausible.Cache.Stats.gather(test)
+    end
+
+    test "get_or_store affects hit rate", %{test: test} do
+      {:ok, _} = start_test_cache(test)
+
+      :ok = ExampleCache.merge_items([{"item1", :item1}], cache_name: test)
+      assert ExampleCache.get("item1", cache_name: test, force?: true)
+
+      # first read of item2 is evaluated from a function and stored
+      assert "value" ==
+               ExampleCache.get_or_store("item2", fn -> "value" end,
+                 cache_name: test,
+                 force?: true
+               )
+
+      # subsequent gets are read from cache and the function is disregarded
+      assert "value" ==
+               ExampleCache.get_or_store("item2", fn -> "disregard" end,
+                 cache_name: test,
+                 force?: true
+               )
+
+      assert "value" ==
+               ExampleCache.get_or_store("item2", fn -> "disregard" end,
+                 cache_name: test,
+                 force?: true
+               )
+
+      # 3 hits, 1 miss
+      assert {:ok, %{hit_rate: 75.0}} = Plausible.Cache.Stats.gather(test)
     end
   end