+++ test/plausible_web/controllers/api/stats_controller/sources_test.exs
@@ -1461,7 +1461,7 @@ defmodule PlausibleWeb.Api.StatsController.SourcesTest do
       ])
 
       conn = get(conn, "/api/stats/#{site.domain}/referrers/Google?period=day")
-      {:ok, terms} = Plausible.Google.Api.Mock.fetch_stats(nil, nil, nil)
+      {:ok, terms} = Plausible.Google.API.Mock.fetch_stats(nil, nil, nil)
 
       assert json_response(conn, 200) == %{
                "total_visitors" => 2,